+++ web/packages/new/photos/types/magicMetadata.ts
@@ -7,18 +7,6 @@ export interface MagicMetadataCore<T> {
 
 export type EncryptedMagicMetadata = MagicMetadataCore<string>;
 
-/**
- * The visibility of an Ente file or collection.
- */
-export enum ItemVisibility {
-    /** The normal state - The item is visible. */
-    visible = 0,
-    /** The item has been archived. */
-    archived = 1,
-    /** The item has been hidden. */
-    hidden = 2,
-}
-
 export enum SUB_TYPE {
     DEFAULT = 0,
     DEFAULT_HIDDEN = 1,
